# src/auth/dto/auth-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';

import { UserResponseDto } from '@users/dto/user-response.dto';

import { TokenResponseDto } from './token-response.dto';

export class AuthResponseDto extends TokenResponseDto {
    @ApiProperty({ type: UserResponseDto })
    user: UserResponseDto;
}