# src/auth/dto/token-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class TokenResponseDto {
    @ApiProperty({
        example:
            'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NiwiaWF0IjoxNzY0MTY5NDkyLCJleHAiOjE3NjQxNjk3OTJ9.Kh8qZxD9zXhP2bU2kjJ8b_RtV2gOoPDr2mPjLq7aB7E'
    })
    accessToken: string;
}