# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/api/users": {
      "get": {
        "operationId": "UsersController_getDtoById",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWrapperResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "get": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "get": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/resend-verification": {
      "get": {
        "operationId": "AuthController_resendVerificationCode",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "info": {
    "title": "Photodel API documentation",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "email": {
            "type": "string",
            "example": "ivan.petrov@example.com"
          },
          "firstName": {
            "type": "string",
            "example": "Иван"
          },
          "lastName": {
            "type": "string",
            "example": "Петров"
          },
          "phone": {
            "type": "string",
            "example": "+7 (912) 345-67-89",
            "nullable": true
          },
          "isAdult": {
            "type": "boolean",
            "example": true
          },
          "isProfessional": {
            "type": "boolean",
            "example": false
          },
          "isVerified": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "example": "2026-02-15T15:31:31.992Z"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "phone",
          "isAdult",
          "isProfessional",
          "isVerified",
          "createdAt"
        ]
      },
      "UserWrapperResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "RegisterRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "firstName": {
            "type": "string",
            "example": "Ivan"
          },
          "lastName": {
            "type": "string",
            "example": "Petrov"
          },
          "isAdult": {
            "type": "boolean",
            "example": true
          },
          "isProfessional": {
            "type": "boolean",
            "example": false
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "isAdult",
          "isProfessional",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NiwiaWF0IjoxNzY0MTY5NDkyLCJleHAiOjE3NjQxNjk3OTJ9.Kh8qZxD9zXhP2bU2kjJ8b_RtV2gOoPDr2mPjLq7aB7E"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NiwiaWF0IjoxNzY0MTY5NDkyLCJleHAiOjE3NjQxNjk3OTJ9.Kh8qZxD9zXhP2bU2kjJ8b_RtV2gOoPDr2mPjLq7aB7E"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "CodeRequestDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "060128"
          }
        },
        "required": [
          "code"
        ]
      }
    }
  }
}